pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-driving-adas-stack"
version = "0.1.0"
description = "Production-style ADAS perception, prediction, and risk streaming stack"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "httpx>=0.27",
  "kagglehub>=1.0.0",
  "numpy>=1.26",
  "pillow>=10.4",
  "pydantic>=2.8",
  "torch>=2.2",
  "torchvision>=0.21",
  "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
include = ["adas_stack*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
